+++ brewery/rpi/element/main.go
@@ -10,7 +10,7 @@ import (
 	"google.golang.org/grpc/reflection"
 )
 
-func StartHeater(port int, pin int) {
+func StartHeater(port int, pin uint8) {
 	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", port))
 	if err != nil {
 		log.Fatalf("failed to listen: %v", err)
